Extend README with a section about CodiMD's role in the fork
In order to provide a fair statement about what people can expect from 
the fork and what not, a little text at the top of the README should 
provide some insight.

Since meta-marked get things done in the way we want, there is no need 
to extend the functionality of it. It does exactly what we want it to do 
and this provides us with all we need.

Making this clear to everyone should help to keep false hopes down.